backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
pinecone==7.3.0
openai==1.97.1
python-dotenv==1.1.1
pydantic-settings==2.10.1
Pillow==11.3.0
numpy==2.3.1

# OCR and Document Processing
pymupdf==1.24.0
paddleocr==2.7.0
easyocr==1.7.0
opencv-python-headless==4.9.0.80
pdf2image==1.17.0

# LangChain for embeddings
langchain-openai==0.1.0
langchain==0.2.0 